:truck: Stores: Moving SimplePersistentActor to new file
