+++ src/main/scala/part3storesandserialization/LocalStores.scala
@@ -1,39 +1,9 @@
 package part3storesandserialization
 
-import akka.actor.{ActorLogging, ActorSystem, Props}
-import akka.persistence.{PersistentActor, RecoveryCompleted, SaveSnapshotFailure, SaveSnapshotSuccess, SnapshotOffer}
+import akka.actor.{ActorSystem, Props}
 import com.typesafe.config.ConfigFactory
 
 object LocalStores extends App {
-
-  class SimplePersistentActor extends PersistentActor with ActorLogging {
-    override def persistenceId: String = "simple-persistent-actor"
-
-    // mutable state
-    var nMessages = 0
-
-    override def receiveCommand: Receive = {
-      case "print" => log.info(s"I have persisted $nMessages so far")
-      case "snap" => saveSnapshot(nMessages)
-      case SaveSnapshotSuccess(metadata) => log.info(s"Save Snapshot was successful: $metadata")
-      case SaveSnapshotFailure(_, cause) => log.warning(s"Save Snapshot failed: $cause")
-      case message => persist(message) { _ =>
-        log.info(s"Persisting $message")
-        nMessages += 1
-      }
-    }
-
-    override def receiveRecover: Receive = {
-      case RecoveryCompleted => log.info(s"Recovery done")
-      case SnapshotOffer(_, payload: Int) =>
-        log.info(s"Recovered snapshot: $payload")
-        nMessages = payload
-      case message =>
-        log.info(s"Recovered: $message")
-        nMessages += 1
-    }
-
-  }
 
   val localStoresActorSystem = ActorSystem("localStoresSystem", ConfigFactory.load().getConfig("localStores"))
   val persistentActor = localStoresActorSystem.actorOf(Props[SimplePersistentActor], "simplePersistentActor")
