+++ src/main/scala/part3storesandserialization/Postgres.scala
@@ -1,5 +1,5 @@
 package part3storesandserialization
 
 object Postgres extends App {
-  
+
 }
