+++ src/main/scala/part3storesandserialization/SimplePersistentActor.scala
@@ -1,0 +1,33 @@
+package part3storesandserialization
+
+import akka.actor.ActorLogging
+import akka.persistence.{PersistentActor, RecoveryCompleted, SaveSnapshotFailure, SaveSnapshotSuccess, SnapshotOffer}
+
+class SimplePersistentActor extends PersistentActor with ActorLogging {
+  override def persistenceId: String = "simple-persistent-actor"
+
+  // mutable state
+  var nMessages = 0
+
+  override def receiveCommand: Receive = {
+    case "print" => log.info(s"I have persisted $nMessages so far")
+    case "snap" => saveSnapshot(nMessages)
+    case SaveSnapshotSuccess(metadata) => log.info(s"Save Snapshot was successful: $metadata")
+    case SaveSnapshotFailure(_, cause) => log.warning(s"Save Snapshot failed: $cause")
+    case message => persist(message) { _ =>
+      log.info(s"Persisting $message")
+      nMessages += 1
+    }
+  }
+
+  override def receiveRecover: Receive = {
+    case RecoveryCompleted => log.info(s"Recovery done")
+    case SnapshotOffer(_, payload: Int) =>
+      log.info(s"Recovered snapshot: $payload")
+      nMessages = payload
+    case message =>
+      log.info(s"Recovered: $message")
+      nMessages += 1
+  }
+
+}
